add http user poll connection

+++ libhttp/source/http-server.c
@@ -104,7 +104,7 @@ const char* http_server_get_header(struct http_session_t *session, const char *n
 int http_server_get_content(struct http_session_t *session, void **content, size_t *length)
 {
 	*content = (void*)http_get_content(session->parser);
-	*length = http_get_content_length(session->parser);
+	*length = (size_t)http_get_content_length(session->parser);
 	return 0;
 }
